+++ docs/history.rst
@@ -5,6 +5,7 @@ Version History
 - Add :func:`sprockets.mixins.mediatype.content.install`.
 - Add :func:`sprockets.mixins.mediatype.content.get_settings`.
 - Deprecate :meth:`sprockets.mixins.mediatype.content.ContentSettings.from_application`.
+- Update to ietfparse 1.4.
 
 `2.1.0`_ (16 Mar 2016)
 ----------------------